Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@23263 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/Utils/ProfilingInfo.cpp
@@ -172,7 +172,7 @@ std::string CProfilingInfo::GetReport() const
         ; iter != end
         ; ++iter)
     {
-        int nCount = (*iter)->GetCount();
+        size_t nCount = (*iter)->GetCount();
         sprintf_s ( lineBuffer, "%7sx %s\n%s:%s\n"
                   , IntToStr (nCount).c_str()
                   , (*iter)->GetName()
